+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-07-22
+### Added
+- New Feature: Support for custom service accounts for Agent Engines during deployment to better adhere to the principle of least privilege.
+- The configured service account for each Agent Engine is now displayed on the Destroy tab.
+
 ## [0.4.0] - 2025-07-08
 ### Added
 - New Feature: Can now list all OAuth Authorizations within the project and can select to delete one